Displays all trainings

+++ Planner4Trainers/Api/Entities/Trainings/Warmups/WarmupOnPlace.cs
@@ -5,7 +5,7 @@
         public int Id { get; set; }
         public string Name { get; set; }
         public string Description { get; set; }
-        public decimal TimeLength { get; set; }
+        public int TimeLength { get; set; }
         public int ExerciseTypeId { get; set; }
         public string ExerciseTypeName { get; set; }
     }
